Fix lint issues in validation tests
